pyproject.toml
[project]
name = "twitter"
version = "0.1.0"
description = "Twitter App"
authors = [
    {name = "caiodiasol",email = "caiodiasol@proton.me"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.4.2,<9.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "pillow (>=12.0.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
